floating point...input 은 최대한 많은 값이 들어온다 생각해야됨...

+++ easy/BOJ_10179.cpp
@@ -5,8 +5,9 @@ int main (void) {
     int T;
     scanf("%d", &T);
     while(T--){
-        float price;
-        scanf("%f", &price);
+        //float price;
+        double price;
+        scanf("%lf", &price);
         printf("$%.2f\n", price * 0.8);
     }
